requirements.txt
bcrypt==4.0.1
click==8.1.3
Flask==2.2.2
Flask-Bcrypt==1.0.1
greenlet==2.0.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
mysqlclient==2.1.1
packaging==23.0
protobuf==3.20.3
python-dotenv==0.21.1
SQLAlchemy==2.0.3
typing_extensions==4.4.0
Werkzeug==2.2.2
